added role admin and user admin isnt restricted

+++ app/Http/Controllers/AdminController.php
@@ -9,18 +9,24 @@ use Illuminate\Http\Request;
 use App\Models\User;
 use App\Models\Post;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 
 class AdminController extends Controller
 {
     public function Statics(){
-        return response()->json([
-            'users_count'=>User::count(),
-            'posts_count'=>Post::count(),
-            'comments_count'=>Comment::count(),]);
+
+        if (auth::user()->role==='admin') {
+            return response()->json([
+                'users_count' => User::count(),
+                'posts_count' => Post::count(),
+                'comments_count' => Comment::count(),]);
+        }
+        return response()->json(['message'=>'Only Admin can access this page'], Response::HTTP_UNAUTHORIZED);
     }
     public function UsersInfo($id){
+        if (auth::user()->role==='admin') {
         $user = User::where('id', $id)->withCount(['posts','comments'])->first();
 
         if (!$user) {
@@ -28,55 +34,66 @@ class AdminController extends Controller
         }
         $user -> makeHidden(['created_at', 'updated_at','phone','role']);
         return response()->json([
-            'users' => $user]);
+            'users' => $user]);}
+        return response()->json(['message'=>'Only Admin can access this page'], Response::HTTP_UNAUTHORIZED);
     }
 
     public function update(Request $request,String $type, $id)
     {
-        $validation = match ($type) {
-            'user'=>app(UpdateUserRequest::class)->validated(),
-            'post'=>app(UpdatePostRequest::class)->validated(),
-            'comment'=>app(UpdateCommentRequest::class)->validated(),
-            default => null
+        if (auth::user()->role==='admin') {
+
+            $validation = match ($type) {
+                'user' => app(UpdateUserRequest::class)->validated(),
+                'post' => app(UpdatePostRequest::class)->validated(),
+                'comment' => app(UpdateCommentRequest::class)->validated(),
+                default => null
             };
-        if (!$validation===null) {
-            return response()->json(['message' => 'Type wasnt found'], Response::HTTP_NOT_FOUND);
-        }
+            if (!$validation === null) {
+                return response()->json(['message' => 'Type wasnt found'], Response::HTTP_NOT_FOUND);
+            }
             $model = match ($type) {
                 'user' => User::class,
                 'post' => Post::class,
                 'comment' => Comment::class,
                 default => null
             };
-        if (!$model===null) {
-            return response()->json(['message' => 'Type wasnt found'], Response::HTTP_NOT_FOUND);
-        }
+            if (!$model === null) {
+                return response()->json(['message' => 'Type wasnt found'], Response::HTTP_NOT_FOUND);
+            }
             $requestType = $model::where('id', $id)->first();
-            if (!$requestType){
-                return response()->json(['message','Type is not found'], Response::HTTP_NOT_FOUND);}
+            if (!$requestType) {
+                return response()->json(['message', 'Type is not found'], Response::HTTP_NOT_FOUND);
+            }
 
-                $requestType->update($validation);
-                return response()->json(["message" => "User updated"], Response::HTTP_ACCEPTED);
+            $requestType->update($validation);
+            return response()->json(["message" => "User updated"], Response::HTTP_ACCEPTED);
+        }
+        return response()->json(['message'=>'Only Admin can access this page'], Response::HTTP_UNAUTHORIZED);
 
         }
         public function delete(Request $request,String $type, $id)
-    {
-            $model = match ($type) {
-                'user' => User::class,
-                'post' => Post::class,
-                'comment' => Comment::class,
-                default => null
-            };
-        if (!$model===null) {
-            return response()->json(['message' => 'Type wasnt found'], Response::HTTP_NOT_FOUND);
-        }
-            $requestType = $model::where('id', $id)->first();
-            if (!$requestType){
-                return response()->json(['message','Type is not found'], Response::HTTP_NOT_FOUND);}
+        {
+            if (auth::user()->role === 'admin') {
+
+                $model = match ($type) {
+                    'user' => User::class,
+                    'post' => Post::class,
+                    'comment' => Comment::class,
+                    default => null
+                };
+                if (!$model === null) {
+                    return response()->json(['message' => 'Type wasnt found'], Response::HTTP_NOT_FOUND);
+                }
+                $requestType = $model::where('id', $id)->first();
+                if (!$requestType) {
+                    return response()->json(['message', 'Type is not found'], Response::HTTP_NOT_FOUND);
+                }
 
                 $requestType->delete();
                 return response()->json(["message" => "the info got deleted"], Response::HTTP_ACCEPTED);
 
+            }
+            return response()->json(['message'=>'Only Admin can access this page'], Response::HTTP_UNAUTHORIZED);
         }
 }
